rename: styles 파일명 변경
