+++ src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import Header from './Header'
 import Footer from './Footer'
-import { Main, ChildrenDiv } from './LayoutStyle'
+import { Main, ChildrenDiv } from './LayoutStyles'
 
 export default function Layout({ children }: { children: JSX.Element }) {
   return (
