+++ src/pages/policy/PolicyStyles.tsx
