+++ src/pages/user/UserDeleteStyles.tsx
