+++ src/pages/user/UserList.tsx
@@ -10,7 +10,7 @@ import {
   SearchUserInput,
   Toast,
   UserListContainer,
-} from './UserListStyle'
+} from './UserListStyles'
 import { useState } from 'react'
 
 export default function UserList() {
